+++ src/Domain/Document/DocumentStatusFacade.php
@@ -32,14 +32,11 @@ final class DocumentStatusFacade extends AbstractAssetFileStatusFacade
         return $assetFile;
     }
 
-    protected function checkDuplicate(AssetFile $assetFile): void
+    protected function checkDuplicate(AssetFile $assetFile): ?AssetFile
     {
-        $originAsset = $this->documentFileRepository->findProcessedByChecksumAndLicence(
+        return $this->documentFileRepository->findProcessedByChecksumAndLicence(
             checksum: $assetFile->getAssetAttributes()->getChecksum(),
             licence: $assetFile->getLicence(),
         );
-        if ($originAsset) {
-            throw new DuplicateAssetFileException($originAsset, $assetFile);
-        }
     }
 }